job-automation/requirements.txt
# Job scraping
feedparser
pandas
requests
beautifulsoup4
lxml

# Email validation & verification
dnspython

# Resume/PDF handling
python-docx
PyMuPDF

# Date/time handling (already in stdlib, but useful)
# datetime - built-in

# SSL handling (already in stdlib)
# ssl - built-in

# IMAP for bounce checking (already in stdlib)
# imaplib - built-in